:sparkles: feature: transaction, data, wallet의 도메인을 각각 정의 #45
