+++ src/main/java/com/automated/trading/stock/StockManager/blockchain/domain/Wallet.java
@@ -1,0 +1,4 @@
+package com.automated.trading.stock.StockManager.blockchain.domain;
+
+public class Wallet {
+}
